Minor bugs in memory size and byte-to-hex formating
There was an off by one error in the memory size, causing 128K machines to trigger the expanded memory tests.

Also the "byte to hex" formatting function was doing it for a word as I'd forgotten to delete one of the translations when I copied and pasted the "word to hex" function.