Fixed serial port output.
Due to (a) documentation of the serial port speed settings for the ZX8302 being wrong, (b) my test for the output buffer being wrong and (c) not having run on a real system, the serial port data wasn't being sent correctly.

I've fixed those issues and, because with the cables I have I can verify it better, I've switched to output to ser2.