Lower memory error output completely re-written
The lower memory test error output routine has been completely re-written.

It now outputs a useful message to the serial port and coloured bands across the screen.

The first is which bank the error was in, white for lower, black for higher.

Then follow 8 bands representing bits 7 to 0, green being correct and red incorrect bits.