External memory size test made more robust & bug in ext memory error report fixed.
It was discovered when testing the ORM on a CST Thor that phantom data above the real RAM was seen by the memory sizing routine as real memory. The test has been change so that it writes and reads two different values and both of these have to pass for the memory address to be valid.

There was also a bug where a faulting memory address had 1 subtracted from it so the address was one too low.