Another attempt at sizing memory correctly on CST Thor
Memory sizing is still not working correctly on the CST Thor.

The theorem I'm working on is that the data bus is floating and doesn't have any pull-ups or anything to drain it so it's retaining the value from the last write. I can't explain that I'm getting valid data from an address I've just written to which isn't real...

Unless the CPU is a 68020 with cache.